requirements.txt
langchain
langchain-openai
gradio
python-dotenv
openai
tiktoken
#ipython
#pydub